+++ src/main/kotlin/us/frollo/frollosdk/extensions/ModelExtension.kt
@@ -62,8 +62,8 @@ internal fun User.updateRequest(): UserUpdateRequest =
         lastName = lastName,
         mobileNumber = mobileNumber,
         gender = gender,
-        address = address,
-        mailingAddress = mailingAddress,
+        address = if (address?.postcode?.isNotBlank() == true) address else null, // This is to avoid the "Invalid postcode length" error from the host
+        mailingAddress = if (mailingAddress?.postcode?.isNotBlank() == true) address else null, // This is to avoid the "Invalid postcode length" error from the host
         householdSize = householdSize,
         householdType = householdType,
         occupation = occupation,